+++ packages/taro-rn-style-transformer/__tests__/config.spec.js
@@ -35,14 +35,14 @@ const defaultConfig = {
 }
 
 async function run (src, filename = './__tests__/styles/a.css', debug) {
-  let options
+  let options = { platform: 'ios' }
   let config
 
   if (typeof src === 'object') {
     ({
       src,
       filename = './__tests__/styles/a.css',
-      options,
+      options = { platform: 'ios' },
       debug,
       config
     } = src || {})
