+++ packages/taro-rn-style-transformer/src/index.ts
@@ -1,22 +1,16 @@
 import * as path from 'path'
 
-// import semver from 'semver'
-// import reactNativePKG from 'react-native/package.json'
 import StyleTransform from './transforms'
+import { Config, TransformOptions } from './types'
 
 const RN_CSS_EXT = ['.css', '.scss', '.sass', '.less', '.styl', '.stylus']
-
-// 目前仅支持React-Native 0.60+
-// const reactNativeVersionString = reactNativePKG.version
-// const reactNativeMinorVersion = semver.minor(reactNativeVersionString)
-
 const upstreamTransformer = require('metro-react-native-babel-transformer')
 
 const getSingleStyleTransform = styleTransformIns()
 
 function styleTransformIns () {
   let styleTransform: StyleTransform | null = null
-  return function (config) {
+  return function (config: Config) {
     // 初始化 config
     if (!styleTransform) {
       styleTransform = new StyleTransform(config)
@@ -25,7 +19,7 @@ function styleTransformIns () {
   }
 }
 
-export async function transform (src: string, filename: string, options) {
+export async function transform (src: string, filename: string, options: TransformOptions) {
   if (typeof src === 'object') {
     // handle RN >= 0.46
     ({ src, filename, options } = src)
@@ -43,13 +37,13 @@ export async function transform (src: string, filename: string, options) {
   return upstreamTransformer.transform({ src, filename, options })
 }
 
-export function rollupTransform (options) {
+export function rollupTransform (options: TransformOptions) {
   return {
     name: 'rn-style-transformer', // this name will show up in warnings and errors
-    async transform (src, filename) {
+    async transform (src: string, filename: string) {
       const ext = path.extname(filename)
       if (RN_CSS_EXT.includes(ext)) {
-        const styleTransform = getSingleStyleTransform((options as any).config)
+        const styleTransform = getSingleStyleTransform(options.config)
         const code = await styleTransform.transform(src, filename, options)
         return { code }
       }
