+++ packages/taro-rn-style-transformer/src/transforms/stylus.ts
@@ -1,111 +1,9 @@
 import stylus from 'stylus'
 
-import { RenderAdditionalResult, RenderResult } from '../types'
+import { RenderAdditionalResult, RenderResult, StylusConfig, StylusRenderOptions } from '../types'
 import { getAdditionalData, insertBefore } from '../utils'
 
-class Evaluator { }
-
-interface Dictionary<T> {
-  [key: string]: T
-}
-
-// https://stylus-lang.com/docs/js.html
-interface RenderOptions {
-  globals?: Dictionary<any>
-  functions?: Dictionary<any>
-  imports?: string[]
-  paths?: string[]
-  filename?: string
-  Evaluator?: typeof Evaluator
-  /**
-   * Specify Stylus plugins to use. Plugins may be passed as
-   * strings instead of importing them in your Webpack config.
-   *
-   * @type {(string|Function)[]}
-   * @default []
-   */
-  use: (string | ((string) => string))[]
-  /**
-   * Add path(s) to the import lookup paths.
-   *
-   * @type {string[]}
-   * @default []
-   */
-  include: string[]
-  /**
-   * Import the specified Stylus files/paths.
-   *
-   * @type {string[]}
-   * @default []
-   */
-  import: string[]
-
-  /**
-   * Define Stylus variables or functions.
-   *
-   * @type {Array|Object}
-   * @default {}
-   */
-  // Array is the recommended syntax: [key, value, raw]
-  define: Array<any> | Record<string, any>
-  // Object is deprecated syntax (there is no possibility to specify "raw')
-  // define: {
-  //   $development: process.env.NODE_ENV === 'development',
-  //   rawVar: 42,
-  // },
-
-  /**
-   * Include regular CSS on @import.
-   *
-   * @type {boolean}
-   * @default false
-   */
-  includeCSS: boolean
-
-  /**
-   * Resolve relative url()'s inside imported files.
-   *
-   * @see https://stylus-lang.com/docs/js.html#stylusresolveroptions
-   *
-   * @type {boolean|Object}
-   * @default { nocheck: true }
-   */
-  // resolveURL: boolean | Record<string, any>,
-  // resolveURL: { nocheck: true },
-
-  /**
-   * Emits comments in the generated CSS indicating the corresponding Stylus line.
-   *
-   * @see https://stylus-lang.com/docs/executable.html
-   *
-   * @type {boolean}
-   * @default false
-   */
-  lineNumbers: boolean
-  /**
-   * @type {boolean}
-   * @default false
-   */
-  disableCache: boolean
-
-  /**
-   * Move @import and @charset to the top.
-   *
-   * @see https://stylus-lang.com/docs/executable.html
-   *
-   * @type {boolean}
-   * @default false
-   */
-  hoistAtrules: boolean
-}
-
-export interface Config {
-  alias?: Record<string, string>
-  options: RenderOptions
-  additionalData?: string | ((key: string) => string)
-}
-
-export const defaultOptions = {
+export const defaultOptions: StylusRenderOptions = {
   use: [],
   include: [],
   import: [],
@@ -117,7 +15,7 @@ export const defaultOptions = {
   disableCache: false
 }
 
-function renderToCSS (src, filename, options = {} as RenderOptions) {
+function renderToCSS (src, filename, options = {} as StylusRenderOptions) {
   const stylusOptions = { filename, ...options }
   const styl = stylus(src, stylusOptions)
 
@@ -212,7 +110,7 @@ function renderToCSS (src, filename, options = {} as RenderOptions) {
   })
 }
 
-export default function transform (src: string, filename: string, config: Config) {
+export default function transform (src: string, filename: string, config: StylusConfig) {
   const additionalData = getAdditionalData(src, config.additionalData)
   const data = insertBefore(src, additionalData)
